# new API log.stats()