--- a/test/stream-pausable.js
+++ b/test/stream-pausable.js
@@ -66,7 +66,7 @@ tape('pausable', function (t) {
 })
 
 tape('close', function (t) {
-  t.equal(log.streams.length, 0, 'no open streams')
+  t.equal(log.streams.size, 0, 'no open streams')
   log.stream({ offsets: false }).pipe({
     paused: false,
     write: function () {},