## Context

Manyverse will need to know how much space _would_ be saved by compaction *before* kickstarting compaction. We don't want the user to reindex leveldb in vain if there are just a handful of deleted records.

## Problem

There is no way of knowing, with the current API, such information, without actually running compaction.

## Solution

`log.stats()` is a new API that does a scan on the log (and this is implemented in AAOL and not in ssb-db2 so to avoid the extra log scan logic it has such as `decrypt`), counting how many deleted records there are and how much space in total those deleted records occupy. This would then be bubbled up to the UI, to the user, so the user can make an informed decision when to compact.

`log.stream(opts)` now also supports `opts.sizes`, to achieve the above.

I also changed `log.streams` from Array to Set, because we don't need the ordering that arrays give us, and `Set.delete` is simpler than splice.