@staltz 2022-07-08
## Log stream benchmark

Here's a quick bench I created to test the performance of `writeToSink` before (master) and after (measure-holes):

```js
const pify = require('util').promisify
const fs = require('fs')
const push = require('push-stream')
const Log = require('.')

const file = '/tmp/aaol-benchmark-stream.log'
try {
  fs.unlinkSync(file)
} catch (_) {}
const log = Log(file, { blockSize: 2 * 1024 })

;(async function () {
  const RECORDS = 1000000
  const records = Array.from({ length: RECORDS }, (_, i) =>
    Buffer.from([i % 256])
  )

  console.time('append ' + RECORDS + ' records')
  await pify(log.append)(records)
  console.timeEnd('append ' + RECORDS + ' records')

  console.time('onDrain')
  await pify(log.onDrain)()
  console.timeEnd('onDrain')

  let i = 0
  await new Promise((resolve) => {
    console.time('stream')
    log.stream({ gt: -1, live: false, offsets: true }).pipe(
      push.drain(
        (record) => {
          i += 1
        },
        () => {
          console.timeEnd('stream');
          resolve()
        }
      )
    )
  })

  if (i !== RECORDS) throw new Error('wrong count: ' + i)

  await pify(log.close)()
})()
```

### Before

```
append 1000000 records: 283.423ms
onDrain: 2315.721ms
stream: 208.770ms
```

```
append 1000000 records: 277.600ms
onDrain: 2082.369ms
stream: 172.664ms
```

```
append 1000000 records: 255.914ms
onDrain: 1757.221ms
stream: 163.800ms
```

### After

```
append 1000000 records: 273.004ms
onDrain: 2281.630ms
stream: 158.369ms
```

```
append 1000000 records: 262.903ms
onDrain: 2013.739ms
stream: 208.999ms
```

```
append 1000000 records: 273.885ms
onDrain: 1662.278ms
stream: 186.753ms
```